+++ braincell/quad/_exp_euler_test.py
@@ -87,7 +87,7 @@ import numpy as np
 
 from braincell.quad.protocol import (
     DiffEqModule,
-    DiffEqState,
+    DiffEqSingleState,
 )
 from braincell.quad import (
     exp_euler_step,
@@ -103,7 +103,7 @@ class _LinearDecay(brainstate.nn.Module, DiffEqModule):
     def __init__(self, x0=1.0, tau_ms=10.0, shape=(3,)):
         super().__init__()
         self.tau = tau_ms * u.ms
-        self.x = DiffEqState(jnp.full(shape, x0, dtype=_FLOAT_DTYPE) * u.mV)
+        self.x = DiffEqSingleState(jnp.full(shape, x0, dtype=_FLOAT_DTYPE) * u.mV)
 
     def compute_derivative(self, *args, **kwargs):
         self.x.derivative = -self.x.value / self.tau